Separate nginx and acme configs.
If nginx is told to use acme it does things like updating the
group of the certificates to allow it to read them. This prevents
the certificates from being used by anything else. Instead add
ngingx to the acme group and allow it to read them that way.

Also attempt to block unknown hosts in nginx instead of it using
and existing host instead.